Add pad function to LMS/ Rework file system
